[Desigh] : Add Category RecyclerView Item v 컬러코드 인덱싱해서 RecyclerView로 RadioBtn 구현 Flexbox 라이브러리 사용, Recyclerview -> FlexboxLayoutManager

+++ app/src/main/java/com/uni/todoary/feature/main/ui/view/MainActivity.kt
@@ -120,7 +120,7 @@ class MainActivity : BaseActivity<ActivityMainBinding>(ActivityMainBinding::infl
         val todolistAdapter = TodoListRVAdapter()
         todolistAdapter.setTodoList(todoList)
         val swipeCallback = TodoListSwipeHelper().apply {
-            setClamp(dpToPx(this@MainActivity, 110f), dpToPx(this@MainActivity, 55f))
+            setClamp(dpToPx(this@MainActivity, 110f).toFloat(), dpToPx(this@MainActivity, 55f).toFloat())
         }
         val swipeHelper = ItemTouchHelper(swipeCallback)
         swipeHelper.attachToRecyclerView(binding.mainSlideTodolistRv)
