+++ app/src/main/java/com/uni/todoary/util/UnitConverter.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import android.util.DisplayMetrics
 import android.util.TypedValue
 
-fun dpToPx(context : Context, dp : Float) : Float{
+fun dpToPx(context : Context, dp : Float) : Int{
     val dm : DisplayMetrics = context.resources.displayMetrics
-    return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp, dm)
+    return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp, dm).toInt()
 }
